Fixes diffHTML tests, some refactoring
Renamed `performance` to `make-measure`, since thats what the function
is. Fixes `process.env` to be a live-lookup from the `process` global if
it exists. Removes more traces of the removed `diff.element` function. A
replacement will be made to support legacy code if folks have it.
