+++ packages/diffhtml/lib/index.js
@@ -93,11 +93,11 @@ import release from './release';
 
 import {
   /**
-   * Adds a global transition listener. With many elements this could be an
+   * Adds a global transition listener. With many DOM Nodes this could be an
    * expensive operation, so try to limit the amount of listeners added if
    * you're concerned about performance.
    *
-   * Since the callback triggers with various elements, most of which you
+   * Since the callback triggers with various DOM Nodes, most of which you
    * probably don't care about, you'll want to filter. A good way of filtering
    * is to use the DOM `matches` method. It's fairly well supported
    * (http://caniuse.com/#feat=matchesselector) and may suit many projects. If
@@ -107,16 +107,16 @@ import {
    *
    *    import { addTransitionState } from 'diffhtml';
    *
-   *    // Fade in all elements as they are added to the DOM.
+   *    // Fade in all Nodes as they are added to the DOM.
    *    addTransitionState('attached', el => $(el).fadeIn().promise());
    *
-   *    // Fade out all elements as they leave the DOM.
+   *    // Fade out all Nodes as they leave the DOM.
    *    addTransitionState('detached', el => $(el).fadeOut().promise());
    *
    *
    * @param state - String name that matches what's available in the
    * documentation above.
-   * @param callback - Function to receive the matching elements.
+   * @param callback - Function to receive the matching DOM Nodes.
    */
   addTransitionState,
 
@@ -199,10 +199,10 @@ const VERSION = '1.0.0-beta';
  * @param {String|Object} markup='' - A string of markup or virtual tree
  * @param {Object =} options={} - An object containing configuration options
  */
-function outerHTML(element, markup='', options={}) {
+function outerHTML(domNode, markup='', options={}) {
   options.inner = false;
-  options.tasks = options.tasks || tasks;
-  return Transaction.create(element, markup, options).start();
+  //options.tasks = options.tasks || tasks;
+  return Transaction.create(domNode, markup, options).start();
 }
 
 /**
@@ -219,14 +219,14 @@ function outerHTML(element, markup='', options={}) {
  *    innerHTML(document.body, 'Hello world');
  *
  *
- * @param {Object} element - A DOM Node to render into
+ * @param {Object} domNode - A DOM Node to render into
  * @param {String|Object} markup='' - A string of markup or virtual tree
  * @param {Object =} options={} - An object containing configuration options
  */
-function innerHTML(element, markup='', options={}) {
+function innerHTML(domNode, markup='', options={}) {
   options.inner = true;
-  options.tasks = options.tasks || tasks;
-  return Transaction.create(element, markup, options).start();
+  //options.tasks = options.tasks || tasks;
+  return Transaction.create(domNode, markup, options).start();
 }
 
 // Public API. Passed to subscribed middleware.
