+++ packages/diffhtml/lib/transaction.js
@@ -1,6 +1,12 @@
+import schedule from './tasks/schedule';
+import shouldUpdate from './tasks/should-update';
+import reconcileTrees from './tasks/reconcile-trees';
+import syncTrees from './tasks/sync-trees';
+import patchNode from './tasks/patch-node';
+import endAsPromise from './tasks/end-as-promise';
 import { StateCache, MiddlewareCache } from './util/caches';
 import { cleanMemory } from './util/memory';
-import makeMeasure from './util/performance';
+import makeMeasure from './util/make-measure';
 import process from './util/process';
 
 export default class Transaction {
