+++ packages/diffhtml/lib/util/process.js
@@ -1,8 +1,6 @@
-const FALLBACK_ENV = 'development';
-const env = typeof process !== 'undefined' ? process.env : {
-  NODE_ENV: FALLBACK_ENV,
+const NODE_ENV = 'development';
+const _process = typeof process !== 'undefined' ? process : {
+  env: { NODE_ENV },
 };
 
-export default {
-  env: { NODE_ENV: env.NODE_ENV || FALLBACK_ENV },
-};
+export default Object.defineProperty({}, 'env', { get: () => _process.env });
